update store with order
